pyproject.toml
[tool.poetry]
name = "taxt-api"
version = "0.1.0"
description = ""
authors = ["polat <polatbekbeknazarov2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.3"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
channels = "^4.0.0"
daphne = "^4.1.0"
channels-redis = "^4.2.0"
celery = "^5.3.6"
djoser = "^2.2.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
django-tailwind = "^3.8.0"
pillow = "^10.3.0"
twisted = {extras = ["http2", "tls"], version = "^24.3.0"}
folium = "^0.17.0"
geopy = "^2.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
